+++ app/Models/Question.php
@@ -4,13 +4,14 @@ namespace App\Models;
 
 use App\Models\Traits\LikeTrait;
 use Backpack\CRUD\app\Models\Traits\CrudTrait;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\MorphMany;
 
 class Question extends Model
 {
-    use CrudTrait, LikeTrait;
+    use CrudTrait, LikeTrait, HasFactory;
 
     /*
     |--------------------------------------------------------------------------