+++ database/seeders/DatabaseSeeder.php
@@ -13,6 +13,8 @@ class DatabaseSeeder extends Seeder
      */
     public function run()
     {
-         \App\Models\User::factory(40)->create();
+        \App\Models\User::factory(20)->create();
+        \App\Models\QuestionCategory::factory(5)->create();
+        \App\Models\Question::factory(5)->create();
     }
 }